rename variables, rename i18nSlice to slice
