+++ packages/mobile/src/i18n/useChangeLanguage.test.tsx
@@ -2,7 +2,7 @@ import { renderHook } from '@testing-library/react-hooks'
 import React from 'react'
 import { Provider } from 'react-redux'
 import i18n from 'src/i18n'
-import * as I18nSlice from 'src/i18n/i18nSlice'
+import * as I18nSlice from 'src/i18n/slice'
 import useChangeLanguage from 'src/i18n/useChangeLanguage'
 import Logger from 'src/utils/Logger'
 import { createMockStore } from 'test/utils'
@@ -12,7 +12,7 @@ jest.mock('src/i18n')
 
 const mockedI18n = mocked(i18n)
 
-const changeLanguageSpy = jest.spyOn(I18nSlice, 'setLanguage')
+const setLanguageSpy = jest.spyOn(I18nSlice, 'setLanguage')
 const loggerErrorSpy = jest.spyOn(Logger, 'error')
 
 const renderHookWithProvider = () => {
@@ -30,8 +30,8 @@ describe('useChangeLanguage', () => {
   })
 
   it('should handle changing language', async () => {
-    const setLanguageSpy = jest.fn().mockResolvedValue(true)
-    mockedI18n.changeLanguage.mockImplementationOnce(setLanguageSpy)
+    const changeLanguageSpy = jest.fn().mockResolvedValue(true)
+    mockedI18n.changeLanguage.mockImplementationOnce(changeLanguageSpy)
 
     const { result } = renderHookWithProvider()
     await result.current('pt-BR')
@@ -41,8 +41,8 @@ describe('useChangeLanguage', () => {
   })
 
   it('should handle error when changing language', async () => {
-    const setLanguageSpy = jest.fn().mockRejectedValue(true)
-    mockedI18n.changeLanguage.mockImplementationOnce(setLanguageSpy)
+    const changeLanguageSpy = jest.fn().mockRejectedValue(true)
+    mockedI18n.changeLanguage.mockImplementationOnce(changeLanguageSpy)
 
     const { result } = renderHookWithProvider()
     await result.current('pt-BR')
