+++ packages/mobile/src/redux/reducers.ts
@@ -11,7 +11,7 @@ import { reducer as fiatExchanges, State as FiatExchangesState } from 'src/fiatE
 import { gethReducer as geth, State as GethState } from 'src/geth/reducer'
 import { reducer as goldToken, State as GoldTokenState } from 'src/goldToken/reducer'
 import { homeReducer as home, State as HomeState } from 'src/home/reducers'
-import i18nReducer, { State as I18nState } from 'src/i18n/i18nSlice'
+import i18nReducer, { State as I18nState } from 'src/i18n/slice'
 import { reducer as identity, State as IdentityState } from 'src/identity/reducer'
 import { reducer as imports, State as ImportState } from 'src/import/reducer'
 import { inviteReducer as invite, State as InviteState } from 'src/invite/reducer'
